move resolvconf parser to dnsContextClient and make it private
