+++ pkg/networkservice/connectioncontext/dnscontext/resolvconf.go
@@ -17,7 +17,7 @@
 // limitations under the License.
 
 // Package resolvconf provides API for editing / reading resolv.conf
-package resolvconf
+package dnscontext
 
 import (
 	"io/ioutil"
@@ -25,15 +25,15 @@ import (
 	"strings"
 )
 
-// ResolveConfig provides API for editing / reading resolv.conf
-type ResolveConfig struct {
+// resolveConfig provides API for editing / reading resolv.conf
+type resolveConfig struct {
 	path       string
 	properties map[string][]string
 }
 
-// OpenResolveConfig reads resolve config file from specific path
-func OpenResolveConfig(p string) (*ResolveConfig, error) {
-	r := &ResolveConfig{
+// openResolveConfig reads resolve config file from specific path
+func openResolveConfig(p string) (*resolveConfig, error) {
+	r := &resolveConfig{
 		path:       p,
 		properties: make(map[string][]string),
 	}
@@ -43,7 +43,7 @@ func OpenResolveConfig(p string) (*ResolveConfig, error) {
 	return r, nil
 }
 
-func (r *ResolveConfig) readProperties() error {
+func (r *resolveConfig) readProperties() error {
 	b, err := ioutil.ReadFile(r.path)
 	if err != nil {
 		return err
@@ -58,12 +58,12 @@ func (r *ResolveConfig) readProperties() error {
 }
 
 // Value returns value of property
-func (r *ResolveConfig) Value(k string) []string {
+func (r *resolveConfig) Value(k string) []string {
 	return r.properties[k]
 }
 
 // SetValue sets value for specific property
-func (r *ResolveConfig) SetValue(k string, values ...string) {
+func (r *resolveConfig) SetValue(k string, values ...string) {
 	if len(values) == 0 {
 		delete(r.properties, k)
 	} else {
@@ -72,7 +72,7 @@ func (r *ResolveConfig) SetValue(k string, values ...string) {
 }
 
 // Save saves resolve config file
-func (r *ResolveConfig) Save() error {
+func (r *resolveConfig) Save() error {
 	var sb strings.Builder
 	var index int
 	for k, v := range r.properties {
@@ -86,12 +86,12 @@ func (r *ResolveConfig) Save() error {
 }
 
 const (
-	// SearchProperty means search list for host-name lookup
-	SearchProperty = "search"
-	// NameserverProperty means name server IP address
-	NameserverProperty = "nameserver"
-	// OptionsProperty  allows certain internal resolver variables to be modified
-	OptionsProperty = "options"
-	// AnyDomain means that allowed any host-name
-	AnyDomain = "."
+	// searchProperty means search list for host-name lookup
+	searchProperty = "search"
+	// nameserverProperty means name server IP address
+	nameserverProperty = "nameserver"
+	// optionsProperty  allows certain internal resolver variables to be modified
+	optionsProperty = "options"
+	// anyDomain means that allowed any host-name
+	anyDomain = "."
 )
